Split out `playtime`
